requirements.txt
requests>=2.25.0
python-dotenv>=0.15.0
websocket-client>=1.3.3
pydantic>=1.10.0
firecrawl>=0.1.0
uuid>=1.30